Addition of a test framework to test the UI
I don't understand everything, and I have big troubles with webpack configuration. Many websites mention a file available at the root of the repo, but we don't have it. So I rolled my own.

Not pretty, and I couldn't get it to work for CSS. I commented out the CSS in the Login.vue for that reason. I wanted the test to compile.